make capatibility table a component
